test: Functions for extracting the synset data as arrays.
The XML converted to an object is nested two layers deep, as far as I could tell. There were many entries, so I chose to convert the objects into flat arrays. Just in a +page.svelte at root of routes. These were fine to look at individually. They basically help me analyze the contents of an XML entry. I don't know XML at all, let alone LMF. Needing to count how many keys of each type there are in the XML, I proceeded to make arrays of arrays of just keys and values within each entry. While I was within that function, I wanted the option of [key[value]] format for the entries instead of a spread. This was achieved, though I fear there might be some lost data due to poor ternary statements and still not knowing all the paramaters used in the XML along with the contents. I found halfway through this round that I totally missed #text, as I didn't expect it. Just checked the docs (quickly) no mention of it, but it has to do with quotes or perhaps definitions from other dicitionaries. Either way, the next step is to study the wordnet and maybe autoencode then validate it on globalwordnet conversion site, just to make sure. These flattened arrays help me get a handle on things, but really the final goal should be to use the object, once I know what I want to grab. I may be avoiding lexical entries altogether, so I'm closer than I think? It would save on load times. I haven't read the paper(s) yet, but as far as I can tell a synset's word is known as a member, definitions are a sentence long and parts of speech are included. A feature for the page would be to change the color of the background when a synset is relative to another word, and divide the background into however many relaive targets there are, each column assigned a number, which may guide key actions, meaning navigation, to traverse the synset array via number keys. Problem is that sometimes relations can nuber greater than 10, forcing perhaps vimmotion like design? Back to back, words are sorted by parts of speech, lexical (?) subgroup and meaning. All in all, a version 1 of this site may just do three things: go fast between the words, let you stop and smell the roses, save your place so you can come back to it + a search feature if you want to get back that way.