feat/style: Functions for WPM measurement for RSVP, member extraction, basic UI, async loading of content.
There are 1.7 times more members than there are synsets. I'd rather the definitions to remain longer since they repeat for every 1.7 members or so. The UI: a progress bar, a space underneath it that says what POS.subgroup is being displayed, which I will be adding a menu button on top of that provides a chance to skip to another 'chapter,' the main display that prints the word in preserved case, the member count (and for my purposes the index, sort of,) and the definition. The page is 600x300, for readability, and responds fine to stretching and zooming. Surely more adjustments could be made to make it more readable. It requires a space for buttons/kb input instructions, jkl I think, and another for form input. I need to allow for searching. Though, not via url query. The display function fills ariables at a given WPM rate. Most likely I'll have to wrap it in another function that feeds it indeces, and place before it a function that takes a word, matches some, and returns possible indices to consider. Things I still want to explore is adding images, a way to suggest the next and previous word in the UI, as surrounding words have similarity in meaning, the site still needs a footer, a logo, share buttons, about modal, a way to store favorite words client side for sessions and bookmarking, and a loading dialogue that is more engaging. The latter will probably be like instrubtions, as the 100 MB wordnet takes a long time to download. Regarding that, maybe there is a server side way accesible to me for dealing with that. Perhaps by preloading the main array or batching it I can cut down on that.