feat: User input and display per index
I neglected to make the display interactable by looping through a nested array of synsets. I flattened out the array making the index of one array a reference to serve the synsets in sequence. Then an automatic increment, with reverse and WPM adjustment. Still needs to handle looping behind 0, handling the 0 index instead of starting on the first member, adding a stop button, and adding a member search. The UI is in test mode, but the inputs all work.