style: SVG loading animation complete.
Since the animation is being done by turning the opacity of each image contained in a stack all of which are set to absolute position, I needed to do responsive CSS for when the screen is tiny. I couldn't get the image to clip at the borders of the viewport. There must be abetter way.